requirements.txt
# Visual RAPTOR ColBERT Integration System - 依存関係

# 基盤フレームワーク
langchain>=0.1.0
langchain-community>=0.0.20
langchain-core>=0.1.0
langchain-ollama>=0.1.0

# 機械学習・深層学習
torch>=2.0.0
transformers>=4.35.0
sentence-transformers>=2.2.0
faiss-cpu>=1.7.4
scikit-learn>=1.3.0
numpy>=1.24.0
pandas>=2.0.0

# 画像処理・OCR
Pillow>=10.0.0
opencv-python>=4.8.0
pytesseract>=0.3.10
easyocr>=1.7.0
paddleocr>=2.7.0

# レイアウト解析
layoutparser>=0.3.4

# 可視化
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.15.0

# データ処理
datasets>=2.14.0
tqdm>=4.65.0
requests>=2.31.0

# ユーティリティ
pathlib2>=2.3.7
python-dotenv>=1.0.0
pydantic>=2.0.0
typing-extensions>=4.7.0

# 開発・テスト
pytest>=7.4.0
pytest-cov>=4.1.0
black>=23.7.0
flake8>=6.0.0
mypy>=1.5.0

# オプション（GPU使用時）
# torch>=2.0.0+cu118  # CUDA 11.8
# faiss-gpu>=1.7.4    # GPU版FAISS

# 日本語処理
fugashi>=1.3.0
ipadic>=1.0.0
mecab-python3>=1.0.6

# 追加のOCRライブラリ
# tesseract  # システムレベルでインストール必要

# ベンチマーク・評価
scikit-metrics>=0.1.0
ir-measures>=0.3.0

# 並列処理
joblib>=1.3.0
multiprocessing-logging>=0.3.4

# ロギング・監視
loguru>=0.7.0
wandb>=0.15.0  # オプション：実験管理

# API・Web（将来的な拡張用）
fastapi>=0.100.0
uvicorn>=0.23.0
streamlit>=1.25.0  # オプション：Webインターフェース